terminada relacion user clientes y reservas

+++ Backend/src/main/java/com/ovniric/controller/ReservationController.java
@@ -1,5 +1,6 @@
 package com.ovniric.controller;
 
+import com.ovniric.dto.ReservationDTO;
 import com.ovniric.model.Reservation;
 import com.ovniric.service.ReservationService;
 import org.springframework.beans.factory.annotation.Autowired;
@@ -11,42 +12,54 @@ import java.util.List;
 import java.util.Optional;
 
 @RestController
-@RequestMapping("/reservations")
+@RequestMapping("/api/reservaciones")
 public class ReservationController {
 
     @Autowired
     private ReservationService reservationService;
 
     @GetMapping
-    public List<Reservation> getAllReservations() {
-        return reservationService.getAllReservations();
+    public ResponseEntity<List<Reservation>> getAllReservations() {
+        List<Reservation> reservations = reservationService.getAllReservations();
+        return new ResponseEntity<>(reservations, HttpStatus.OK);
     }
 
     @GetMapping("/{id}")
-    public ResponseEntity<Reservation> getReservationById(@PathVariable Long id) {
+    public ResponseEntity<Reservation> getReservationById(@PathVariable("id") Long id) {
         Optional<Reservation> reservation = reservationService.getReservationById(id);
         if (reservation.isPresent()) {
-            return ResponseEntity.ok(reservation.get());
+            return new ResponseEntity<>(reservation.get(), HttpStatus.OK);
         } else {
-            return ResponseEntity.notFound().build();
+            return new ResponseEntity<>(HttpStatus.NOT_FOUND);
         }
     }
 
     @PostMapping
-    public ResponseEntity<Reservation> createReservation(@RequestBody Reservation reservation) {
-        Reservation createdReservation = reservationService.createReservation(reservation);
-        return ResponseEntity.status(HttpStatus.CREATED).body(createdReservation);
+    public ResponseEntity<Reservation> createReservation(@RequestBody ReservationDTO reservationDTO) {
+        Reservation reservation = reservationService.createReservation(reservationService.toReservation(reservationDTO));
+        return new ResponseEntity<>(reservation, HttpStatus.CREATED);
     }
 
-    @PutMapping()
-    public ResponseEntity<Reservation> updateReservation(@RequestBody Reservation reservation) {
-        Reservation updatedReservation = reservationService.updateReservation(reservation);
-        return ResponseEntity.ok(updatedReservation);
+    @PutMapping
+    public ResponseEntity<Reservation> updateReservation(@RequestBody Reservation updatedReservation) {
+        Optional<Reservation> reservation = reservationService.getReservationById(updatedReservation.getId());
+        if (reservation.isPresent()) {
+            updatedReservation.setId(updatedReservation.getId());
+            Reservation newReservation = reservationService.updateReservation(updatedReservation);
+            return new ResponseEntity<>(newReservation, HttpStatus.OK);
+        } else {
+            return new ResponseEntity<>(HttpStatus.NOT_FOUND);
+        }
     }
 
     @DeleteMapping("/{id}")
-    public ResponseEntity<Void> deleteReservation(@PathVariable Long id) {
-        reservationService.deleteReservation(id);
-        return ResponseEntity.noContent().build();
+    public ResponseEntity<Void> deleteReservation(@PathVariable("id") Long id) {
+        Optional<Reservation> reservation = reservationService.getReservationById(id);
+        if (reservation.isPresent()) {
+            reservationService.deleteReservation(id);
+            return new ResponseEntity<>(HttpStatus.NO_CONTENT);
+        } else {
+            return new ResponseEntity<>(HttpStatus.NOT_FOUND);
+        }
     }
 }
